# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * รินจัดให้ - Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /** @minLength 2 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface ServiceInput {
  name: string;
  description: string;
  icon: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  icon: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface Package {
  id: number;
  serviceId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  durationDays: number;
  features: string[];
  isPopular: boolean;
  createdAt: string;
}

export interface ServiceWithPackages {
  id: number;
  name: string;
  description: string;
  icon: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
  packages: Package[];
}

export interface PackageInput {
  name: string;
  description?: string;
  price: number;
  durationDays: number;
  features?: string[];
  isPopular?: boolean;
}

export interface OrderInput {
  packageId: number;
  note?: string;
  discordUsername?: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
  processing: 'processing',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  adminNote?: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  processing: 'processing',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  userId: number;
  packageId: number;
  status: OrderStatus;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  adminNote?: string | null;
  /** @nullable */
  discordUsername?: string | null;
  totalPrice?: number;
  createdAt: string;
  user?: User;
  package?: Package;
}

export interface SlipInput {
  orderId: number;
  slipImageBase64: string;
  amount?: number;
}

export type PaymentVerifyInputStatus = typeof PaymentVerifyInputStatus[keyof typeof PaymentVerifyInputStatus];


export const PaymentVerifyInputStatus = {
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface PaymentVerifyInput {
  status: PaymentVerifyInputStatus;
  note?: string;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface Payment {
  id: number;
  orderId: number;
  slipImageBase64: string;
  /** @nullable */
  amount?: number | null;
  status: PaymentStatus;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  verifiedAt?: string | null;
  createdAt: string;
}

export interface GalleryImageInput {
  imageBase64: string;
  title: string;
}

export interface GalleryImage {
  id: number;
  imageBase64: string;
  title: string;
  createdAt: string;
}

export interface DashboardStats {
  totalOrders: number;
  pendingOrders: number;
  completedOrders: number;
  totalRevenue: number;
  totalUsers: number;
  recentOrders: Order[];
}