# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: รินจัดให้ - Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── SERVICES ────────────────────────────────────────────────
...
  /services/{serviceId}/packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List packages for a service
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a package (admin)
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    put:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ORDERS ──────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── PAYMENTS ─────────────────────────────────────────────────
...
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery images
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: uploadGalleryImage
      tags: [gallery]
      summary: Upload gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete gallery image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ADMIN DASHBOARD ─────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ─── AUTH ───────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ─── SERVICES ────────────────────────────────────────────────
...
    PackageInput:
      type: object
      required: [name, price, durationDays]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationDays:
          type: integer
        features:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
    Package:
      type: object
      required: [id, serviceId, name, price, durationDays, features, isPopular, createdAt]
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationDays:
          type: integer
        features:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── ORDERS ──────────────────────────────────────────────────
    OrderInput:
      type: object
      required: [packageId]
      properties:
        packageId:
          type: integer
        note:
          type: string
        discordUsername:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, paid, processing, completed, cancelled]
        adminNote:
          type: string
    Order:
      type: object
      required: [id, userId, packageId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        packageId:
          type: integer
        status:
          type: string
          enum: [pending, paid, processing, completed, cancelled]
        note:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        discordUsername:
          type: ["string", "null"]
        totalPrice:
          type: number
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        package:
          $ref: "#/components/schemas/Package"

    # ─── PAYMENTS ─────────────────────────────────────────────────
...
    GalleryImageInput:
      type: object
      required: [imageBase64, title]
      properties:
        imageBase64:
          type: string
        title:
          type: string
    GalleryImage:
      type: object
      required: [id, imageBase64, title, createdAt]
      properties:
        id:
          type: integer
        imageBase64:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time

    # ─── ADMIN DASHBOARD ─────────────────────────────────────────
...