# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * รินจัดให้ - Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';
import type { Package } from './package';
import type { User } from './user';

export interface Order {
  id: number;
  userId: number;
  packageId: number;
  status: OrderStatus;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  adminNote?: string | null;
  /** @nullable */
  discordUsername?: string | null;
  totalPrice?: number;
  createdAt: Date;
  user?: User;
  package?: Package;
}