# lib/api-zod/src/generated/types/packageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * รินจัดให้ - Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageInput {
  name: string;
  description?: string;
  price: number;
  durationDays: number;
  features?: string[];
  isPopular?: boolean;
}